tests/requirements.txt
pytest==8.2.2
pytest-django==4.8.0
beautifulsoup4==4.12.3